backend/requirements.txt
langchain-groq
langgraph
langchain
pypdf
reportlab
python-dotenv